# Conflicts:
#	ivycpg/src/main/java/com/ivy/cpg/view/order/OrderSummary.java
#	ivycpg/src/main/res/layout/dialog_order_remarks.xml

+++ ivycpg/src/main/java/com/ivy/sd/png/provider/SynchronizationHelper.java
@@ -1,7 +1,6 @@
 package com.ivy.sd.png.provider;
 
 import android.app.Activity;
-import android.content.ContentValues;
 import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
@@ -17,7 +16,6 @@ import android.text.TextUtils;
 import android.util.Base64;
 import android.widget.Toast;
 
-import com.amazonaws.services.s3.model.S3ObjectSummary;
 import com.amazonaws.util.StringInputStream;
 import com.android.volley.AuthFailureError;
 import com.android.volley.DefaultRetryPolicy;
@@ -65,7 +63,6 @@ import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
-import java.io.FileReader;
 import java.io.FilenameFilter;
 import java.io.IOException;
 import java.io.InputStream;
@@ -134,6 +131,7 @@ SynchronizationHelper {
 
     public static final String USER_ID = "UserId";
     public static final String VERSION_CODE = "VersionCode";
+    public static final String VERSION_NAME = "VersionName";
     public static final String MOBILE_DATE_TIME = "MobileDateTime";
     public static final String MOBILE_UTC_DATE_TIME = "MobileUTCDateTime";
     public static final String REQUEST_MOBILE_DATE_TIME = "RequestDate";
@@ -1131,13 +1129,14 @@ SynchronizationHelper {
 
     public void deleteFiles(String folderPath, String fnamesStarts) {
         File folder = new File(folderPath);
+        if (folder.listFiles() != null) {
+            File files[] = folder.listFiles();
 
-        File files[] = folder.listFiles();
-
-        for (File tempFile : files) {
-            if (tempFile != null) {
-                if (tempFile.getName().startsWith(fnamesStarts))
-                    tempFile.delete();
+            for (File tempFile : files) {
+                if (tempFile != null) {
+                    if (tempFile.getName().startsWith(fnamesStarts))
+                        tempFile.delete();
+                }
             }
         }
     }
@@ -1371,6 +1370,7 @@ SynchronizationHelper {
             json.put("UserId", bmodel.userMasterHelper.getUserMasterBO()
                     .getUserid());
             json.put("VersionCode", bmodel.getApplicationVersionNumber());
+            json.put(SynchronizationHelper.VERSION_NAME, bmodel.getApplicationVersionName());
             JSONArray jsonArray = new JSONArray();
             for (RetailerMasterBO retailerMasterBO : retailerList) {
                 jsonArray.put(retailerMasterBO.getRetailerID());
@@ -1406,6 +1406,7 @@ SynchronizationHelper {
                 json.put("UserId", bmodel.userMasterHelper.getUserMasterBO()
                         .getUserid());
                 json.put("VersionCode", bmodel.getApplicationVersionNumber());
+                json.put(SynchronizationHelper.VERSION_NAME, bmodel.getApplicationVersionName());
 
                 int insert = VOLLEY_DOWNLOAD_INSERT;
                 if (whichDownload == DownloadType.RETAILER_WISE_DOWNLOAD) {
@@ -1438,6 +1439,7 @@ SynchronizationHelper {
                 json.put("UserId", bmodel.userMasterHelper.getUserMasterBO()
                         .getUserid());
                 json.put("VersionCode", bmodel.getApplicationVersionNumber());
+                json.put(SynchronizationHelper.VERSION_NAME, bmodel.getApplicationVersionName());
                 if (whichDownload == 1)
                     json.put("IsRetailer", 1);
                 else if (whichDownload == 2)
@@ -1911,6 +1913,7 @@ SynchronizationHelper {
             json.put("UserId", bmodel.userMasterHelper.getUserMasterBO()
                     .getUserid());
             json.put("VersionCode", bmodel.getApplicationVersionNumber());
+            json.put(SynchronizationHelper.VERSION_NAME, bmodel.getApplicationVersionName());
             json.put("MobileDate", Utils.getDate("yyyy/MM/dd HH:mm:ss"));
             json.put("MobileUTCDateTime",
                     Utils.getGMTDateTime("yyyy/MM/dd HH:mm:ss"));
@@ -2415,6 +2418,7 @@ SynchronizationHelper {
             jsonObj.put("Model", Build.MODEL);
             jsonObj.put("VersionCode",
                     bmodel.getApplicationVersionNumber());
+            jsonObj.put(SynchronizationHelper.VERSION_NAME,bmodel.getApplicationVersionName());
             jsonObj.put("DeviceId",
                     bmodel.activationHelper.getIMEINumber());
             jsonObj.put("RegistrationId", bmodel.regid);
@@ -2468,6 +2472,7 @@ SynchronizationHelper {
             jsonObj.put("Model", Build.MODEL);
             jsonObj.put("VersionCode",
                     bmodel.getApplicationVersionNumber());
+            jsonObj.put(SynchronizationHelper.VERSION_NAME,bmodel.getApplicationVersionName());
             jsonObj.put("DeviceId",
                     bmodel.activationHelper.getIMEINumber());
             jsonObj.put("RegistrationId", bmodel.regid);
@@ -2749,6 +2754,7 @@ SynchronizationHelper {
                 json.put("UserId", bmodel.userMasterHelper.getUserMasterBO()
                         .getUserid());
                 json.put("VersionCode", bmodel.getApplicationVersionNumber());
+                json.put(SynchronizationHelper.VERSION_NAME, bmodel.getApplicationVersionName());
                 json.put("RetailerId", retailerid);
                 int size = mNewRetailerDownloadUrlList.size();
                 for (String url : mNewRetailerDownloadUrlList) {
@@ -2809,6 +2815,7 @@ SynchronizationHelper {
                         .getUserid());
                 json.put("UserId", bmodel.retailerMasterBO.getSelectedUserID());
                 json.put("VersionCode", bmodel.getApplicationVersionNumber());
+                json.put(SynchronizationHelper.VERSION_NAME, bmodel.getApplicationVersionName());
                 json.put("RetailerId", retailerid);
                 int size = mUserRetailerTranDownloadUrlList.size();
                 for (String url : mUserRetailerTranDownloadUrlList) {
@@ -2905,6 +2912,7 @@ SynchronizationHelper {
             json.put("UserId", bmodel.userMasterHelper.getUserMasterBO()
                     .getUserid());
             json.put("VersionCode", bmodel.getApplicationVersionNumber());
+            json.put(SynchronizationHelper.VERSION_NAME, bmodel.getApplicationVersionName());
 
             callVolley(DataMembers.SERVER_URL + appendSihUrl, fromWhere, 1, SIH_DOWNLOAD, json);
         } catch (Exception e) {
@@ -2934,6 +2942,7 @@ SynchronizationHelper {
             json.put("UserId", bmodel.userMasterHelper.getUserMasterBO()
                     .getUserid());
             json.put("VersionCode", bmodel.getApplicationVersionNumber());
+            json.put(SynchronizationHelper.VERSION_NAME, bmodel.getApplicationVersionName());
 
 
             downloadurl = DataMembers.SERVER_URL + downloadurl;
@@ -2979,6 +2988,7 @@ SynchronizationHelper {
                 json.put("UserId", bmodel.userMasterHelper.getUserMasterBO()
                         .getUserid());
                 json.put("VersionCode", bmodel.getApplicationVersionNumber());
+                json.put(SynchronizationHelper.VERSION_NAME, bmodel.getApplicationVersionName());
                 JSONArray jsonArray = new JSONArray();
                 jsonArray.put(id);
                 json.put("LocationIds", jsonArray);
@@ -2986,6 +2996,7 @@ SynchronizationHelper {
                 json.put("LoginUserId", bmodel.userMasterHelper.getUserMasterBO().getUserid());
                 json.put("UserId", id);
                 json.put("VersionCode", bmodel.getApplicationVersionNumber());
+                json.put(SynchronizationHelper.VERSION_NAME, bmodel.getApplicationVersionName());
             }
 
 
@@ -3027,6 +3038,7 @@ SynchronizationHelper {
             json.put("UserId", bmodel.userMasterHelper.getUserMasterBO()
                     .getUserid());
             json.put("VersionCode", bmodel.getApplicationVersionNumber());
+            json.put(SynchronizationHelper.VERSION_NAME, bmodel.getApplicationVersionName());
             json.put("Key", retailerName);
 
 
@@ -3051,6 +3063,7 @@ SynchronizationHelper {
             json.put("UserId", bmodel.userMasterHelper.getUserMasterBO()
                     .getUserid());
             json.put("VersionCode", bmodel.getApplicationVersionNumber());
+            json.put(SynchronizationHelper.VERSION_NAME, bmodel.getApplicationVersionName());
 
 
             callVolley(sb.toString(), fromWhere, 0, updateWhere, json);
@@ -3425,6 +3438,7 @@ SynchronizationHelper {
                         bmodel.activationHelper.getIMEINumber());
                 jsonFormatter.addParameter("VersionCode",
                         bmodel.getApplicationVersionNumber());
+                jsonFormatter.addParameter(SynchronizationHelper.VERSION_NAME,bmodel.getApplicationVersionName());
                 jsonFormatter.addParameter("OrganisationId", bmodel.userMasterHelper
                         .getUserMasterBO().getOrganizationId());
                 if (isDayClosed()) {
@@ -3886,6 +3900,7 @@ SynchronizationHelper {
                         bmodel.activationHelper.getIMEINumber());
                 jsonFormatter.addParameter("VersionCode",
                         bmodel.getApplicationVersionNumber());
+                jsonFormatter.addParameter(SynchronizationHelper.VERSION_NAME,bmodel.getApplicationVersionName());
                 jsonFormatter.addParameter("OrganisationId", bmodel.userMasterHelper
                         .getUserMasterBO().getOrganizationId());
                 jsonFormatter.addParameter("MobileDate",
@@ -4014,6 +4029,7 @@ SynchronizationHelper {
                         bmodel.activationHelper.getIMEINumber());
                 jsonFormatter.addParameter("VersionCode",
                         bmodel.getApplicationVersionNumber());
+                jsonFormatter.addParameter(SynchronizationHelper.VERSION_NAME,bmodel.getApplicationVersionName());
                 jsonFormatter.addParameter("OrganisationId", bmodel.userMasterHelper
                         .getUserMasterBO().getOrganizationId());
                 if (isDayClosed()) {
@@ -4126,6 +4142,7 @@ SynchronizationHelper {
             json.put("UserId", bmodel.userMasterHelper.getUserMasterBO()
                     .getUserid());
             json.put("VersionCode", bmodel.getApplicationVersionNumber());
+            json.put(SynchronizationHelper.VERSION_NAME, bmodel.getApplicationVersionName());
             json.put("PhoneNo", phoneNum);
 
             callVolley(sb.toString(), FROM_SCREEN.COUNTER_SALES_SELECTION, 1, VOLLEY_CUSTOMER_SEARCH, json);
@@ -4148,6 +4165,7 @@ SynchronizationHelper {
             jsonObject.put("UserId", bmodel.userMasterHelper.getUserMasterBO()
                     .getUserid());
             jsonObject.put("VersionCode", bmodel.getApplicationVersionNumber());
+            jsonObject.put(SynchronizationHelper.VERSION_NAME, bmodel.getApplicationVersionName());
             for (TeamLeadBO teamLeadBO : absenteesList) {
                 if (teamLeadBO.getAttendance() == 0) {
 
@@ -4216,6 +4234,7 @@ SynchronizationHelper {
                         .getUserid());
                 jsonObj.put("LoginId", bmodel.userMasterHelper.getUserMasterBO().getLoginName());
                 jsonObj.put("VersionCode", bmodel.getApplicationVersionNumber());
+                jsonObj.put(SynchronizationHelper.VERSION_NAME, bmodel.getApplicationVersionName());
                 http.setParamsJsonObject(jsonObj);
 
                 http.connectMe();
@@ -4533,6 +4552,7 @@ SynchronizationHelper {
                     .getUserMasterBO().getLoginName());
             jsonFormatter.addParameter("VersionCode",
                     bmodel.getApplicationVersionNumber());
+            jsonFormatter.addParameter(SynchronizationHelper.VERSION_NAME,bmodel.getApplicationVersionName());
             jsonFormatter.addParameter("DistributorId", bmodel.userMasterHelper
                     .getUserMasterBO().getDistributorid());
             jsonFormatter.addParameter("OrganisationId", bmodel.userMasterHelper
@@ -4854,6 +4874,7 @@ SynchronizationHelper {
             json.put(SynchronizationHelper.USER_ID, bmodel.userMasterHelper.getUserMasterBO()
                     .getUserid());
             json.put(SynchronizationHelper.VERSION_CODE, bmodel.getApplicationVersionNumber());
+            json.put(SynchronizationHelper.VERSION_NAME, bmodel.getApplicationVersionName());
             json.put(SynchronizationHelper.MOBILE_DATE_TIME,
                     Utils.getDate("yyyy/MM/dd HH:mm:ss"));
             json.put(SynchronizationHelper.MOBILE_UTC_DATE_TIME,
